requirements.txt
#pyhcl==0.4.5
python-hcl2==4.3.2
requests==2.31.0
pydantic==2.6.3
pymongo==4.6.2
PyGithub==2.2.0
beanie==1.25.0
# torpy==1.1.6
# scholarly==1.7.11